feat(global): hidden sidebar
