+++ client/src/components/Global/InfoIcon/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import st from 'styles/samples/info_icon.module.css'
+import st from 'styles/components/info_icon.module.css'
 
 export const InfoIcon = ({onClick=()=>{}}) => 
   <div className={st.info_icon} onClick={onClick}>i</div>
