+++ client/src/components/UserSpace/CompanyCard/EmpRow.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import { usePush, useHttp } from 'hooks'
-import { TextInputFrame } from 'samples/Global/TextInputFrame'
-import { GenerateIcon } from 'samples/Global/GenerateIcon'
-import { ToggleSwitch } from 'samples/Global/ToggleSwitch'
-import { InfoIcon } from 'samples/Global/InfoIcon'
+import { TextInputFrame } from 'components/Global/TextInputFrame'
+import { GenerateIcon } from 'components/Global/GenerateIcon'
+import { ToggleSwitch } from 'components/Global/ToggleSwitch'
+import { InfoIcon } from 'components/Global/InfoIcon'
 import { TrashIcon } from '../TrashIcon'
 import { generatePassword, convertDate } from 'utils'
 import { useMyContext } from 'Context'
