+++ client/src/components/UserSpace/CompanyCard/index.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useState, useCallback } from 'react'
 import { useHttp, usePush } from 'hooks'
-import { RadioButton } from 'samples/UserSpace/RadioButton'
-import { ToggleSwitch } from 'samples/Global/ToggleSwitch'
-import { TrashIcon } from 'samples/UserSpace/TrashIcon'
+import { RadioButton } from 'components/UserSpace/RadioButton'
+import { ToggleSwitch } from 'components/Global/ToggleSwitch'
+import { TrashIcon } from 'components/UserSpace/TrashIcon'
 import { EmpRow } from './EmpRow'
 import { convertDate } from 'utils/convertDate'
 import { generatePassword } from 'utils/generatePass'
