+++ client/src/hooks/push.hook.js
@@ -1,6 +1,6 @@
 import { useCallback } from 'react'
 import { toast } from 'materialize-css'
-import c from '../styles/samples/push.module.css'
+import c from '../styles/components/push.module.css'
 
 export const usePush = () => useCallback((messages, ok=false) => {
     if (messages){
