+++ client/src/styles/UserSpace/main_styles.css
@@ -1,15 +1,20 @@
 section{
-    display: flex;
+  display: flex;
 }
 
 header{
-    width: 200px;
-    height: 100vh;
-    padding: 20px 30px;
-    border-right: 1px solid rgba(25, 52, 77, .1);
+  position: fixed;
+  top: 0;
+  overflow: hidden;
+  width: 200px;
+  height: 100vh;
+  border-right: 1px solid rgba(25, 52, 77, .1);
+  z-index: 10;
+  background: #fff;
+  transition: all .2s ease-in-out;
 }
 
 main{
-    width: calc(100% - 200px);
-    height: 100vh;
+  width: 100%;
+  height: 100vh;
 }
