+++ client/src/styles/UserSpace/navlink.module.css
@@ -1,4 +1,5 @@
 .passive_navlink, .active_navlink{
+    margin: 25px 30px;
     text-decoration: none;
     cursor: pointer;
     user-select: none;
