+++ client/src/styles/components/generate_icon.module.css
