+++ client/src/styles/components/toggle_switch.module.css
