+++ client/src/styles/components/trash_icon.module.css
